+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2026-05-18
+
+### Changed
+- add pal_research with 7 discrete palettes and 6 gradient scales; update plot_manhattan colors to coral/teal
+- Updated files: R/academic_color_scales.R,R/plot_manhattan.R,NAMESPACE,man
+
+
 ## [1.14.1] - 2026-04-27
 
 ### Changed